# internal/app/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rebecca API",
    "version": "0.3.0",
    "description": "Go-native Rebecca panel API for dashboard, bot, and third-party client developers.\n\nAuthentication: the dashboard uses an opaque server-side session stored in the HttpOnly `rebecca_session` cookie. Session permissions are loaded from the database on every request, so permission changes take effect without logging out. Bots and external integrations should use personal `rk_...` API keys as `Authorization: Bearer <api_key>`. Legacy JWT creation at `POST /admin/token` and `POST /api/admin/token` remains available only for older Marzban-compatible clients.\n\nDo not automate username/password login for bots. API keys can be revoked independently and inherit the owner admin permissions, service scope, and limits.\n\nMutation behavior: user, service, node, config, and host mutations are transactional where applicable. Runtime changes are synchronized to nodes through gRPC and recorded for recovery."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Rebecca panel origin"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/admin": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get current admin",
        "description": "Return the authenticated admin context, role, permissions, limits, and service assignments.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getAdmin"
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create admin",
        "description": "Create a new admin if the current admin can manage admins.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postAdmin"
      }
    },
    "/api/admins": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List admins",
        "description": "List non-deleted admins with optional username filtering, pagination, and sorting.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "name": "username",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getAdmins"
      }
    },
    "/api/admin/{username}": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Modify admin",
        "description": "Modify an existing admin's profile, role, permissions, limits, and service links.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Username"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "putAdminUsername"
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete admin",
        "description": "Soft-delete an admin and handle owned users according to the request behavior.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Username"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "deleteAdminUsername"
      }
    },
...
    "/api/v2/services/{service_id}/usage/admins": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "Service usage by admin",
        "description": "Return service usage distribution across admins.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Requires an authenticated admin. Listing/detail respects service visibility. Service mutation, admin limits, reset, and auto-inbound actions require sudo/full_access.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ServiceID"
          },
          {
            "$ref": "#/components/parameters/Start"
          },
          {
            "$ref": "#/components/parameters/End"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getV2ServicesServiceIdUsageAdmins"
      }
    },
...
    "/api/settings/subscriptions/admins/{admin_id}": {
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update admin subscription override",
        "description": "Update per-admin subscription domain and subscription settings override.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Panel read requires an authenticated admin. Settings mutation, backup, templates, and TODO/410 settings routes require sudo/full_access.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "parameters": [
          {
            "$ref": "#/components/parameters/AdminID"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "putSettingsSubscriptionsAdminsAdminId"
      }
    },
...
    "/api/auth/session": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get dashboard session",
        "description": "Resolve the opaque dashboard cookie, enforce idle and absolute expiry, reload the current admin and permissions from the database, and return the current login state. The dashboard polls this endpoint so permission changes are reflected without a new login.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current server-side login state and fresh admin permission context.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSession"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "getAuthSession"
      }
    },
...
    "/api/auth/2fa/setup": {
      "post": {
        "tags": [
          "My Account"
        ],
        "summary": "Start two-factor enrollment",
        "description": "Generate a new TOTP secret for the current admin and keep it pending on the current session until confirmation. The secret is returned once as text and an otpauth URI; it is not enabled until a valid code is submitted.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "One-time authenticator enrollment secret and otpauth URI.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TOTPSetup"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postAuth2faSetup"
      }
    },
    "/api/auth/2fa/confirm": {
      "post": {
        "tags": [
          "My Account"
        ],
        "summary": "Confirm two-factor enrollment",
        "description": "Validate a code from the pending TOTP secret, enable two-factor authentication for the current admin, activate a setup-required session, and revoke every other active login session for that admin.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/OTPCode"
        },
        "responses": {
          "200": {
            "description": "Current server-side login state and fresh admin permission context.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSession"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postAuth2faConfirm"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Rebecca admin API key or legacy JWT",
        "description": "Use a personal rk_ API key for bots and third-party integrations. Legacy JWTs from /admin/token remain accepted for older Marzban-compatible clients."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rebecca_session",
        "description": "Opaque HttpOnly dashboard session cookie issued by POST /api/auth/login. It is managed by the browser and is not readable by dashboard JavaScript."
      }
    },
    "parameters": {
...
      "Offset": {
        "name": "offset",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 0
        },
        "description": "Zero-based row offset used for paginated list endpoints."
      },
...
    },
...
  }
}